Modularity improvements (#170)
* Moved methods around to increase modularity

* Moved CUDA-specific methods out of the CUDA library
* Moved Packet-Specific methods into packet library
* Removed uses of static variables (happened as part of moving)
* Added bsg_manycore_tile.cpp/.h to replace bsg_manycore_epa.h (same functionality)
* Renamed hb_mc_set_tile_group_origin to hb_mc_tile_set_group_origin
