+++ cl_manycore/libraries/bsg_manycore_elf.cpp
@@ -1,16 +1,26 @@
+#include <stdlib.h>
+#include <string.h>
+#include <stdio.h>
+#include <stdint.h>
+#include <unistd.h>
+#include <inttypes.h>
+#include <elf.h>
 #ifndef COSIM
-#include <bsg_manycore_elf.h>
+#include <bsg_manycore_driver.h>
 #include <bsg_manycore_errno.h>
+#include <bsg_manycore_mem.h>
+#include <bsg_manycore_elf.h>
 #else
-#include "bsg_manycore_elf.h"
+#include "bsg_manycore_driver.h"
 #include "bsg_manycore_errno.h"
+#include "bsg_manycore_mem.h"
+#include "bsg_manycore_elf.h"
 #endif
+
 
 #include <map>
 #include <string>
 #include <memory>
-
-
 
 using std::string;
 using std::map;
