+++ cl_manycore/libraries/bsg_manycore_elf.h
@@ -4,20 +4,12 @@
 #ifdef __cplusplus 
 extern "C" {
 #endif
-#include <stdlib.h>
-#include <string.h>
-#include <unistd.h>
-#include <stdio.h>
-#include <getopt.h>
-#include <stdint.h>
-#include <alloca.h>
-#include <inttypes.h>
-#include <elf.h>
 #ifndef COSIM
-#include <bsg_manycore_driver.h>
+#include <bsg_manycore_mem.h>
 #else
-#include "bsg_manycore_driver.h"
+#include "bsg_manycore_mem.h"
 #endif
+
 int symbol_to_eva(const char *fname, const char *sym_name, eva_t *eva);
 #ifdef __cplusplus
 }
