+++ cl_manycore/libraries/bsg_manycore_mem.h
@@ -20,8 +20,21 @@
 extern "C" {
 #endif
 
+typedef uint32_t eva_id_t;
+
+typedef uint32_t eva_t;
+
+typedef struct {
+	uint32_t x;
+	uint32_t y; 
+	uint32_t epa;
+} npa_t;
+
 int hb_mc_copy_from_epa (uint8_t fd, hb_mc_response_packet_t *buf, uint32_t x, uint32_t y, uint32_t epa, uint32_t size); 
 int hb_mc_copy_to_epa (uint8_t fd, uint32_t x, uint32_t y, uint32_t epa, uint32_t *buf, uint32_t size);
+int hb_mc_eva_to_npa (eva_id_t eva_id, eva_t eva, npa_t *npa);
+int hb_mc_npa_to_eva (eva_id_t eva_id, npa_t *npa, eva_t *eva);
+
 
 #ifdef __cplusplus
 }
