+++ cl_manycore/regression/spmd/spmd_tests.h
@@ -6,6 +6,7 @@
 #include <utils/sh_dpi_tasks.h>
 #include "fpga_pci_sv.h"
 #include "bsg_manycore_driver.h"
+#include "bsg_manycore_tile.h"
 #include "bsg_manycore_mem.h"
 #include "bsg_manycore_loader.h"
 #include "bsg_manycore_errno.h"	
@@ -13,6 +14,7 @@
 #else // !COSIM
 
 #include <bsg_manycore_driver.h>
+#include <bsg_manycore_tile.h>
 #include <bsg_manycore_mem.h>
 #include <bsg_manycore_loader.h>
 #include <bsg_manycore_errno.h>
